requirements.txt
streamlit>=1.22
langchain>=0.0.250
openai>=0.27.0
python-dotenv
sqlalchemy
pydantic
pytest
